refactor(capabilities): migrate the http server cap to the struct-hosted actor macro
Move the http server `impl NativeActor` into runtime.rs under `#[runtime]`
and put `#[actor(singleton)]` on the `HttpServerCapability` struct (#2385).
